[refactor]: modularize consultation draft generator functionality
