+++ backend/aspnet-core/src/MedStream.Application/Services/ConsultationService/ConsultationDraftGenerator.OpenAi.cs
@@ -1,0 +1,82 @@
+using Newtonsoft.Json;
+using System;
+using System.Net.Http;
+using System.Net.Http.Headers;
+using System.Text;
+using System.Text.Json;
+using System.Threading.Tasks;
+
+namespace MedStream.Consultation;
+
+public partial class ConsultationDraftGenerator
+{
+    private async Task<JsonElement?> TryGenerateJsonAsync(string systemPrompt, object payload)
+    {
+        var apiKey = _configuration["OpenAI:ApiKey"] ?? Environment.GetEnvironmentVariable("OPENAI_API_KEY");
+        if (string.IsNullOrWhiteSpace(apiKey))
+        {
+            return null;
+        }
+
+        var endpoint = (_configuration["OpenAI:BaseUrl"] ?? "https://api.openai.com/v1").TrimEnd('/') + "/chat/completions";
+        var model = _configuration["OpenAI:Model"] ?? "gpt-4o-mini";
+
+        using var httpClient = new HttpClient();
+        httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", apiKey);
+
+        var requestBody = new
+        {
+            model,
+            temperature = 0.2,
+            response_format = new { type = "json_object" },
+            messages = new object[]
+            {
+                new { role = "system", content = systemPrompt },
+                new { role = "user", content = JsonConvert.SerializeObject(payload) }
+            }
+        };
+
+        try
+        {
+            using var content = new StringContent(JsonConvert.SerializeObject(requestBody), Encoding.UTF8, "application/json");
+            using var response = await httpClient.PostAsync(endpoint, content);
+            if (!response.IsSuccessStatusCode)
+            {
+                return null;
+            }
+
+            var responseJson = await response.Content.ReadAsStringAsync();
+            using var responseDocument = JsonDocument.Parse(responseJson);
+            var completionContent = responseDocument.RootElement
+                .GetProperty("choices")[0]
+                .GetProperty("message")
+                .GetProperty("content")
+                .GetString();
+            if (string.IsNullOrWhiteSpace(completionContent))
+            {
+                return null;
+            }
+
+            using var contentDocument = JsonDocument.Parse(completionContent);
+            return contentDocument.RootElement.Clone();
+        }
+        catch
+        {
+            return null;
+        }
+    }
+
+    private static string ReadString(JsonElement element, string propertyName, string fallbackValue)
+    {
+        if (element.TryGetProperty(propertyName, out var property) && property.ValueKind == JsonValueKind.String)
+        {
+            var value = property.GetString()?.Trim();
+            if (!string.IsNullOrWhiteSpace(value))
+            {
+                return value;
+            }
+        }
+
+        return fallbackValue;
+    }
+}
